Edit and delete to be perfomed by owner

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
@@ -92,13 +93,21 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session, RedirectAttributes redirectAttributes) {
         Image image = imageService.getImage(imageId);
 
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        if(isOwner(imageId, session)) {
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
+        else {
+            String title = imageService.getImage(imageId).getTitle();
+            String errMessage = "Only the owner of the image can edit the image";
+            redirectAttributes.addAttribute("editError", errMessage).addFlashAttribute("editError", errMessage);
+            return "redirect:/images/" + imageId + '/' + title;
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -140,9 +149,17 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, RedirectAttributes redirectAttributes) {
+        if(isOwner(imageId, session)) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else {
+            String title = imageService.getImage(imageId).getTitle();
+            String errMessage = "Only the owner of the image can edit the image";
+            redirectAttributes.addAttribute("deleteError", errMessage).addFlashAttribute("deleteError", errMessage);
+            return "redirect:/images/" + imageId + '/' + title;
+        }
     }
 
 
@@ -186,5 +203,17 @@ public class ImageController {
         tagString.append(lastTag.getName());
 
         return tagString.toString();
+    }
+
+    //The method to check whether logged in user is actual owner of uploaded image
+    //If user is actual owner of image, method returns true, else it returns false
+    private boolean isOwner(Integer imageId, HttpSession session) {
+        Image image = imageService.getImage(imageId);
+        Integer userId = image.getUser().getId();
+
+        User loggedUser = (User) session.getAttribute("loggeduser");
+        Integer loggedUserId = loggedUser.getId();
+
+        return userId == loggedUserId ? true: false;
     }
 }
